add preview card spinner
