+++ src/lib/constants.ts
@@ -1,0 +1,1 @@
+export const PAGE_SIZE = 15
